backend/requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
pymongo==4.3.3
dnspython==2.3.0
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.5
email-validator==2.0.0
bcrypt==4.0.1
opencv-python==4.7.0.72
face-recognition==1.3.0
numpy==1.24.2
pydantic==1.10.7
dlib==19.24.0
python-dotenv==1.0.0 